more

+++ app/models/Patient.scala
@@ -4,10 +4,10 @@ import scala.slick.driver.H2Driver.simple._
 import play.api.db.slick.DB
 import play.api.Play.current
 
-case class Patient(id: Option[Int] = None, name: String, age: Int)
+case class Patient(id: Option[Long] = None, name: String, age: Int)
 
 object Patients extends Table[Patient]("patient") {
-   def id = column[Int]("id", O.PrimaryKey, O.AutoInc)
+   def id = column[Long]("id", O.PrimaryKey, O.AutoInc)
    def name = column[String]("name")
    def age = column[Int]("age")
    def * = id.? ~ name ~ age <>(Patient, Patient.unapply _)
@@ -22,7 +22,7 @@ object Patients extends Table[Patient]("patient") {
      (for { c <- Patients } yield c).list
    }
 
-   def findById(id: Int): Option[Patient] = DB.withSession { implicit session =>
-       Patients.byId(id).firstOption
+   def findById(id: Long): Option[Patient] = DB.withSession { implicit session =>
+     Patients.byId(id).firstOption
    }
 }
